requirements.txt
aiohappyeyeballs==2.6.1
aiohttp~=3.8.5
aiosignal~=1.3.1
attrs~=25.1.0
certifi~=2024.8.30
charset-normalizer~=2.1.1
datasets==3.5.0
dill==0.3.8
filelock~=3.12.2
frozenlist~=1.4.0
fsspec~=2024.10.0
huggingface-hub~=0.26.2
idna~=3.4
Jinja2~=3.1.2
MarkupSafe~=2.1.1
mpmath==1.3.0
multidict~=6.0.4
multiprocess==0.70.16
networkx==3.4.2
numpy~=1.26.4
packaging==24.2
pandas==2.2.3
propcache==0.3.1
pyarrow==19.0.1
python-dateutil~=2.8.2
pytz~=2022.6
PyYAML~=6.0
regex~=2024.9.11
requests==2.32.3
safetensors~=0.4.5
six~=1.16.0
sympy==1.13.1
tokenizers~=0.21.0
torch==2.6.0
tqdm==4.67.1
transformers~=4.47.0
typing_extensions~=4.12.2
tzdata~=2025.1
urllib3~=1.26.13
xxhash==3.5.0
yarl~=1.9.2

SQLAlchemy~=2.0.36
Faker~=37.1.0
beautifulsoup4~=4.13.3
markdownify~=1.1.0
RapidFuzz~=3.12.2
jupyterlab-pygments~=0.2.2
pip~=25.0.1
distro~=1.9.0
pyOpenSSL~=25.0.0
cryptography~=44.0.0
pillow~=10.4.0
rich~=13.9.4
colorama~=0.4.6
ipython~=8.7.0
notebook~=6.5.2
pytest~=7.3.1
setuptools~=65.5.0
cffi~=1.15.1
mypy~=1.3.0
decorator~=5.1.1
psutil~=6.1.1
lxml~=5.3.0
elasticsearch~=8.8.0
arrow~=1.2.3
Pygments~=2.13.0
nltk~=3.9.1
fastapi~=0.115.8
pydantic~=2.10.6
starlette~=0.45.3
uvicorn~=0.34.0
openai~=1.61.1
sentence-transformers~=4.0.2
faiss-cpu~=1.10.0
scipy~=1.15.2
scikit-learn~=1.6.1